refactor: 모바일 메뉴-모달 코드 분리
